Refactor: move JetpackAITranscriptionResponse out of rest client and adjust import statements
