+++ fluxc/src/main/java/org/wordpress/android/fluxc/network/rest/wpcom/jetpackai/JetpackAITranscriptionRestClient.kt
@@ -175,33 +175,6 @@ class JetpackAITranscriptionRestClient  @Inject constructor(
         }
     }
 
-    enum class JetpackAITranscriptionErrorType {
-        API_ERROR,
-        AUTH_ERROR,
-        GENERIC_ERROR,
-        INVALID_RESPONSE,
-        TIMEOUT,
-        NETWORK_ERROR,
-        CONNECTION_ERROR,
-        // local errors
-        INELIGIBLE_AUDIO_FILE,
-        PARSE_ERROR,
-        // HTTP
-        BAD_REQUEST,
-        NOT_FOUND,
-        NOT_AUTHENTICATED,
-        REQUEST_TOO_LARGE,
-        SERVER_ERROR
-    }
-
-    sealed class JetpackAITranscriptionResponse {
-        data class Success(val model: String) : JetpackAITranscriptionResponse()
-        data class Error(
-            val type: JetpackAITranscriptionErrorType,
-            val message: String? = null
-        ) : JetpackAITranscriptionResponse()
-    }
-
     @Suppress("MagicNumber")
     private fun fromHttpStatusCode(code: Int): JetpackAITranscriptionErrorType {
         return when (code) {
